Added some commands

+++ src/Handler/RemoveVoteHandler.php
@@ -1,18 +1,15 @@
 <?php
-/**
- * This file is part of the ceo.
+
+/*
+ * This file is part of the ceo project.
  *
  * (c) Aula de Software Libre de la UCO <aulasoftwarelibre@uco.es>
- * (c) Sergio Gómez <sergio@uco.es>
- * (c) Omar Sotillo <i32sofro@uco.es>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-
 namespace App\Handler;
-
 
 use App\Command\RemoveVoteCommand;
 use App\Entity\Vote;
@@ -27,6 +24,7 @@ class RemoveVoteHandler
 
     /**
      * RemoveVoteHandler constructor.
+     *
      * @param VoteRepository $repository
      */
     public function __construct(VoteRepository $repository)
@@ -37,6 +35,7 @@ class RemoveVoteHandler
     public function handle(RemoveVoteCommand $command)
     {
         $idea = $command->getIdea();
+        $owner = $idea->getOwner();
         $user = $command->getUser();
 
         $vote = $this->repository->findOneBy([
@@ -44,7 +43,7 @@ class RemoveVoteHandler
             'idea' => $idea,
         ]);
 
-        if ($vote instanceof Vote) {
+        if ($vote instanceof Vote && false === $owner->equalsTo($user)) {
             $this->repository->remove($vote);
         }
     }
